Move actitivies to dependence package
