+++ app/src/main/java/mnxk/kotlintex/projectm/activities/SignInActivity.kt
@@ -1,7 +1,8 @@
-package mnxk.kotlintex.projectm
+package mnxk.kotlintex.projectm.activities
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import mnxk.kotlintex.projectm.R
 import mnxk.kotlintex.projectm.databinding.ActivitySignInBinding
 
 class SignInActivity : AppCompatActivity() {
